+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Journal des modifications
 9.4.0
 -----
 - Passage à PHP 8.2.
+- Module Saisie Thèse : Renseignement de la discipline SISE rendue obligatoire
+- Accès donné aux directeurs d'ED à la page "dépôt fichier divers"
+- Module Formation : Les doctorants peuvent maintenant se réinscrire à une session de formation dans de nouveaux cas : 
+  - Si la première session est clôturée, et que le doctorant est non classé ou en liste complémentaire
 - [FIX] Page des comptes utilisateurs : erreur 'Type of Application\Controller\UtilisateurController::$shibService must be UnicaenAuthentification\Service\ShibService'.
 - [FIX] Envoi vers STEP-STAR : pas de balise INE du tout dans le fichier TEF si l'INE du doctorant est vide.